requirements.txt
asgiref==3.9.1
Django==5.2.4
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
environs==14.2.0
marshmallow==4.0.0
pillow==11.3.0
PyJWT==2.9.0
python-dotenv==1.1.1
sqlparse==0.5.3
